HAZOP using RAG/requirements.txt
# requirements.txt
# List of Python packages required to run the HAZOP automation scripts.
# Install these packages using pip:
# pip install -r requirements.txt

# For connecting to and interacting with the Neo4j database.
neo4j

# For handling data in tabular format, especially for CSVs and Excel generation.
pandas

# For creating and manipulating the graph structure in memory before loading to Neo4j.
networkx

# For parsing the DEXPI XML file.
lxml

# For generating formatted Excel files. Works with pandas.
xlsxwriter

# For making API calls to the Large Language Model (e.g., OpenAI).
openai

# For creating vector embeddings for semantic search.
sentence-transformers

# For efficient similarity search in the RAG step.
faiss-cpu

# For progress bars to monitor long-running processes.
tqdm

# For parsing RDF/XML ontology files.
rdflib
